backend/requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-jsonform==2.22.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-writable-nested==0.7.0
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1